% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_analytics.R
\name{brew_analytics}
\alias{brew_analytics}
\title{Retrieve analytics data for a formula or cask}
\usage{
brew_analytics(
  package,
  days = 30,
  category = c("install", "install-on-request", "build-error"),
  ...
)
}
\arguments{
\item{package}{Name of one or more formulae or casks}

\item{days}{Number of days' worth of analytics data to retrieve. One of 30
(default), 90, or 365.}

\item{category}{Type of analytics data to retrieve. One of \code{"install"}
(default), \code{"install-on-request"}, or \code{"build-error"}.}

\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\description{
Retrieve analytics data for a formula or cask
}
\examples{
# Retrieve 90-day install data for unixodbc
brew_analytics("unixodbc", days = 90)
}
