% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_cleanup}
\alias{brew_cleanup}
\title{Clean up Homebrew environment}
\usage{
brew_cleanup(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\description{
\code{brew_cleanup()} removes stale lock files and outdated downloads
for all formulae and casks, and removes old versions of installed formulae.
}
