% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_config.R
\name{brew_config}
\alias{brew_config}
\title{Show Homebrew and System Configuration}
\usage{
brew_config()
}
\value{
A string containing the raw text output from \verb{brew config}
}
\description{
\code{brew_config()} returns important information about your Homebrew
installation and other related system information.
}
\examples{
\dontrun{
brew_config()
}
}
