% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_help}
\alias{brew_help}
\title{Get help}
\usage{
brew_help(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\value{
A string
}
\description{
\code{brew_help()} can be used to get information about Homebrew or
the available commands
}
\examples{
\dontrun{
# Get help about the 'install' command
brew_help("install")
}
}
