% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_info}
\alias{brew_info}
\title{Display information about TODO}
\usage{
brew_info(package = NULL, ...)
}
\arguments{
\item{package}{Name of formula or cask. If one is not given, info about your
homebrew environment is shown.}

\item{...}{
  Arguments passed on to \code{\link[=brew_cleanup]{brew_cleanup}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{brew_info()} displays brief information about your Homebrew
installation or specific packages.
}
\examples{
# Get info about Homebrew installation
brew_info()

# Get info about unixodbc package
brew_info("unixodbc")
}
