% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_list.R
\name{brew_list}
\alias{brew_list}
\title{List all installed formulae and casks}
\usage{
brew_list(formulae = TRUE, casks = TRUE, ...)
}
\arguments{
\item{formulae}{Include installed formulae (default: \code{TRUE})}

\item{casks}{Include installed casks (default: \code{TRUE})}
}
\value{
A data frame containing the \code{name} and \code{version} of each installed
package and how it was installed (i.e., via formula or cask).
}
\description{
List all installed formulae and casks
}
\examples{
brew_list()
}
