% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_outdated}
\alias{brew_outdated}
\title{List outdated packages}
\usage{
brew_outdated(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\value{
TODO
}
\description{
\code{brew_outdated()} returns a list of installed packages that have
an updated version available
}
