% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_update}
\alias{brew_update}
\title{Update Homebrew environment and formulae}
\usage{
brew_update(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cleanup]{brew_cleanup}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{brew_update()} fetches the newest version of Homebrew and all
formulae
}
