% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_upgrade}
\alias{brew_upgrade}
\title{Upgrade outdated packages}
\usage{
brew_upgrade(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cleanup]{brew_cleanup}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{brew_upgrade()} upgrades outdated casks and outdated, unpinned
formulae using the same options they were originally installed with, plus
any appended brew formula options.
}
