% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_version.R
\name{brew_version}
\alias{brew_version}
\title{Get Homebrew version information}
\usage{
brew_version()
}
\value{
A named list with items \code{homebrew}, \code{core}, and \code{cask}
}
\description{
\code{brew_version()} returns the version numbers of Homebrew,
Homebrew/homebrew-core and Homebrew/homebrew-cask (if tapped)
}
\examples{
brew_version()
}
