% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_homebrew.R
\name{has_homebrew}
\alias{has_homebrew}
\title{Check whether Homebrew is installed}
\usage{
has_homebrew()
}
\value{
A logical value indicating whether Homebrew could be found
}
\description{
\code{has_homebrew()} checks whether Homebrew is installed on this
machine. Although most functions in this package allow the path to Homebrew's
\code{brew} command to be specified, this function only checks whether \code{brew}
exists in the system's path.
}
\examples{
has_homebrew()
}
\seealso{
\url{https://brew.sh} for installation instructions
}
