% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homebrew-package.R
\docType{package}
\name{homebrew-package}
\alias{homebrew}
\alias{homebrew-package}
\title{homebrew: Install and Manage Packages using 'Homebrew'}
\description{
Manage packages on macOS and Linux machines using the 'Homebrew' package management system <https://brew.sh>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/homebrew}
  \item Report bugs at \url{https://github.com/briandconnelly/homebrew/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID})

}
\keyword{internal}
