% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed}
\alias{is_installed}
\title{Determine whether a package is installed}
\usage{
is_installed(package, formulae = TRUE, casks = TRUE)
}
\arguments{
\item{package}{Name of formula or cask}

\item{formulae}{Include installed formulae (default: \code{TRUE})}

\item{casks}{Include installed casks (default: \code{TRUE})}
}
\value{
A logical value indicating whether or not the package is installed
}
\description{
\code{is_installed()} checks if a given package is installed either
via formula or cask
}
\examples{
# Check whether or not unixodbc is installed
is_installed("unixodbc")
}
