% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_cmd.R
\name{brew_cmd}
\alias{brew_cmd}
\title{Run a homebrew command}
\usage{
brew_cmd(
  cmd,
  args = character(),
  brew_cmd = Sys.which("brew"),
  echo = rlang::is_interactive(),
  ...
)
}
\arguments{
\item{cmd}{The command to run (e.g., \code{install})}

\item{args}{Additional arguments for the command}

\item{brew_cmd}{Optional: Location of \code{brew} command.}

\item{echo}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}

\item{...}{Additional arguments passed to \code{\link[processx:run]{processx::run()}}}
}
\value{
A logical value indicating whether the command was successful or not
}
\description{
Run a homebrew command
}
\examples{
# TODO
}
