% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{brew_doctor}
\alias{brew_doctor}
\title{TODO}
\usage{
brew_doctor(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cleanup]{brew_cleanup}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{brew_doctor()} TODO
}
