% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_info.R
\name{brew_info}
\alias{brew_info}
\title{Display information about a formula or cask}
\usage{
brew_info(package, ...)
}
\arguments{
\item{package}{Name of one or more formulae or casks}

\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\value{
A list of lists. Each sub-list contains information about a package
or cask specified by the \code{package} argument.
}
\description{
\code{brew_info()} displays detailed information about a formula or
cask.
}
\note{
System and package analytics can be retrieved with \code{\link{brew_analytics}}
}
\examples{
# Get info about unixodbc package
brew_info("unixodbc")
}
