% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_install.R
\name{brew_install}
\alias{brew_install}
\alias{brew_install_formula}
\alias{brew_install_cask}
\alias{brew_reinstall}
\alias{brew_uninstall}
\title{Install a formula or cask}
\usage{
brew_install(
  package,
  cask = FALSE,
  force = FALSE,
  only_dependencies = FALSE,
  keep_tmp = FALSE,
  ...
)

brew_install_formula(package, ...)

brew_install_cask(package, ...)

brew_reinstall(package, ...)

brew_uninstall(package, ...)
}
\arguments{
\item{package}{Name of formula or cask}

\item{cask}{If \code{TRUE}, \code{package} refers to a cask. If \code{FALSE} (default),
\code{package} refers to a formula.}

\item{force}{Do not check for previous installations
(default: \code{FALSE})}

\item{only_dependencies}{Install only dependencies and not the formula itself
(default: \code{FALSE})}

\item{keep_tmp}{Do not delete temporary files created during installation
(default: \code{FALSE})}

\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\value{
A logical value indicating whether or not the installation was a
success
}
\description{
\code{brew_install()} installs a formula or cask.
For information about additional arguments, see \code{brew_help("install")}.

\code{brew_reinstall()} uninstalls and then reinstalls a formula or
cask using the same options it was originally installed with, plus any
appended options specific to a formula.

\code{brew_uninstall()} uninstalls a formula or cask
}
\examples{
\dontrun{
# Install unixodbc
brew_install("unixodbc")
}
\dontrun{
# Install the Homebrew-compiled version of R
brew_install_formula("r")
}
\dontrun{
# Install RStudio cask
brew_install_cask("rstudio")
}
\dontrun{
# Reinstall unixodbc
brew_reinstall("unixodbc")
}
\dontrun{
# Uninstall unixodbc
brew_uninstall("unixodbc")
}
}
