% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_install.R
\name{brew_install}
\alias{brew_install}
\alias{brew_install_formula}
\alias{brew_install_cask}
\title{Install a formula or cask}
\usage{
brew_install(
  x,
  cask = FALSE,
  force = FALSE,
  only_dependencies = FALSE,
  keep_tmp = FALSE,
  ...
)

brew_install_formula(x, ...)

brew_install_cask(x, ...)
}
\arguments{
\item{x}{Name(s) of formula or cask}

\item{cask}{If \code{TRUE}, \code{x} refers to a cask. If \code{FALSE} (default), \code{x} refers to a formula.}

\item{force}{Install without checking for previous installations (default: \code{FALSE})}

\item{only_dependencies}{Install only dependencies and not the formula itself (default: \code{FALSE})}

\item{keep_tmp}{Do not delete temporary files created during installation (default: \code{FALSE})}

\item{...}{Optional arguments for \code{\link[=brew_cmd]{brew_cmd()}}}
}
\value{
A logical value indicating whether or not the installation was a success
}
\description{
Install a formula or cask
}
\examples{
library(homebrew)

\dontrun{
brew_install("unixodbc")}
}
