% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_list.R
\name{brew_list}
\alias{brew_list}
\title{List all installed formulae and casks}
\usage{
brew_list(formulae = TRUE, casks = TRUE)
}
\arguments{
\item{formulae}{Whether to search installed formulae (default: \code{TRUE})}

\item{casks}{Whether to search installed casks (default: \code{TRUE})}
}
\value{

}
\description{
\code{brew_list()} returns a data frame containing the names and versions of all
installed formulae and casks.
}
