% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_list.R
\name{brew_list}
\alias{brew_list}
\alias{is_installed}
\title{Check installed formulae and casks}
\usage{
brew_list(formulae = TRUE, casks = TRUE, ...)

is_installed(package, formulae = TRUE, casks = TRUE)
}
\arguments{
\item{formulae}{Include installed formulae (default: \code{TRUE})}

\item{casks}{Include installed casks (default: \code{TRUE})}

\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}

\item{package}{Name of formula or cask}
}
\value{
A data frame containing the \code{name} and
\code{version} of each installed package and how it was installed (i.e., via
formula or cask).

A logical value indicating whether or not the package is installed
}
\description{
\code{brew_list()} lists all installed packages

\code{is_installed()} checks if a given package is installed
}
\examples{
\dontrun{
brew_list()
}
\dontrun{
# Check whether or not unixodbc is installed
is_installed("unixodbc")
}
}
