% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_version.R
\name{brew_version}
\alias{brew_version}
\title{Get Homebrew version information}
\usage{
brew_version(...)
}
\value{
A named list containing the following items:

\describe{
\item{\code{homebrew}}{String describing the version of Homebrew}
\item{\code{core}}{String describing git commit info for
\href{https://github.com/Homebrew/homebrew-core}{core formulae}}
\item{\code{cask}}{String describing git commit info for
\href{https://github.com/Homebrew/homebrew-cask}{casks}, if tapped}
}
}
\description{
\code{brew_version()} returns version info about Homebrew, core formulae, and
casks.
}
\note{
Version information for installed packages can be retrieved with
\code{\link[=brew_list]{brew_list()}}
}
\examples{
brew_version()
}
