% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew_version.R
\name{brew_version}
\alias{brew_version}
\title{Get Homebrew version information}
\usage{
brew_version(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=brew_cmd]{brew_cmd}}
  \describe{
    \item{\code{cmd}}{The command to run (e.g., \code{install})}
    \item{\code{args}}{Additional arguments for the command}
    \item{\code{brew_cmd}}{Optional: Location of \code{brew} command.}
    \item{\code{echo}}{Whether to print output and error messages to the screen.
Defaults to \code{TRUE} for interactive sessions.}
  }}
}
\value{
A named list containing the following items:

\describe{
\item{\code{homebrew}}{String describing the version of Homebrew}
\item{\code{core}}{String describing git commit info for
\href{https://github.com/Homebrew/homebrew-core}{core formulae}}
\item{\code{cask}}{String describing git commit info for
\href{https://github.com/Homebrew/homebrew-cask}{casks}, if tapped}
}
}
\description{
\code{brew_version()} returns version info about Homebrew, core formulae, and
casks.
}
\note{
Version information for installed packages can be retrieved with
\code{\link[=brew_list]{brew_list()}}
}
\examples{
brew_version()
}
