% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed}
\alias{is_installed}
\alias{is_formula_installed}
\alias{is_cask_installed}
\title{Determine whether a formula or cask is installed}
\usage{
is_installed(x, formulae = TRUE, casks = TRUE)

is_formula_installed(x)

is_cask_installed(x)
}
\arguments{
\item{x}{Name(s) of formula or cask}

\item{formulae}{Search installed formulae (default: \code{TRUE})}

\item{casks}{Search installed casks (default: \code{TRUE})}
}
\value{

}
\description{
\code{is_formula_installed()} TODO

\code{is_cask_installed()} TODO
}
\examples{
\dontrun{
library(homebrew)

is_installed("unixodbc")
}
}
